progress: 6.34% of code matched (601/8216 functions, 540/935 objects)

+++ src/game/game_fn_8002FAE8.c
@@ -36,7 +36,7 @@ extern int fn_801CEB2C(); extern int fn_801D1B10(); extern int fn_801D38BC();
 extern int fn_801D38E8(); extern int fn_801D39E0(); extern int fn_801DD0A8();
 extern int fn_801E2B28(); extern int fn_801E8328(); extern int fn_802006D4();
 extern int fn_80200C10(); extern int fn_80200C38(); extern int fn_8020104C();
-extern int fn_8020123C(); extern int fn_80201814(); extern int fn_80201AE4();
+extern unsigned long long fn_8020123C(); extern int fn_80201814(); extern int fn_80201AE4();
 extern int fn_80201B44(); extern int fn_80201B54(); extern int fn_80201B5C();
 extern int fn_80201B8C(); extern int fn_80201B9C(); extern int fn_80201BC0();
 extern int fn_80201BC8(); extern int fn_80201D14(); extern int fn_80201D1C();
